requirements.txt
fastapi==0.104.1
google-generativeai==0.3.0
python-dotenv==1.0.0
supabase==2.0.3
uvicorn==0.24.0
mlflow>=2.11.0
wandb>=0.16.3
streamlit>=1.32.0
pandas>=2.2.0
plotly>=5.19.0